Refactor unit tests.
